Config sync changes
