+++ resources/modinfo.json
@@ -6,7 +6,7 @@
     "Professor Cupcake"
   ],
   "description": "Changes all combat damage to bleeding",
-  "version": "0.0.10",
+  "version": "0.0.11",
   "dependencies": {
     "game": "1.20.0-rc.8"
   },
